requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
streamlit==1.28.0
python-dotenv==1.0.0
openpyxl==3.1.2
pyarrow==14.0.1